+++ mizani/tests/test_formatters.py
@@ -100,23 +100,15 @@ def test_log_format():
     assert formatter([1, 35, 60, 1e6]) == ['1', '4e1', '6e1', '1e6']
 
     formatter = log_format(base=2)
-    assert formatter([1, 2, 4, 8]) == ['1', '10', '100', '1000']
-    assert formatter([0b1, 0b10, 0b11, 0b1011]) == ['1', '10', '11', '1011']
+    assert formatter([1, 2, 4, 8]) == ['2^0', '2^1', '2^2', '2^3']
+    assert formatter([0b1, 0b10, 0b11]) == ['2^0', '2^1', '2^1.585']
 
     formatter = log_format(base=8)
-    assert formatter([1, 4, 8, 16, 64]) == ['1', '4', '10', '20', '100']
+    assert formatter([1, 4, 8, 64]) == ['8^0', '8^0.667', '8^1', '8^2']
 
-    formatter = log_format(base=16)
-    assert formatter([1, 8, 16, 32, 256]) == ['1', '8', '10', '20', '100']
-
-    # Floats
-    formatter = log_format(base=8)
-    assert formatter([1., 4., 8., 16.]) == ['1', '4', '10', '20']
-
-    # Fallback to base 10
     formatter = log_format(base=np.e)
-    with pytest.warns(UserWarning, match=r"base = 10$"):
-        assert formatter([1, 8, 16, 32]) == ['1', '8', '16', '32']
+    assert formatter([1, np.pi, np.e**2, np.e**3]) == \
+        ['e^0', 'e^1.145', 'e^2', 'e^3']
 
     # mathtex
     formatter = log_format(mathtex=True)
@@ -124,6 +116,9 @@ def test_log_format():
         ['$10^{-3}$', '$10^{-1}$', '$10^{4}$']
     assert formatter([35, 60]) == ['35', '60']
     assert formatter([1, 10000]) == ['$10^{0}$', '$10^{4}$']
+
+    formatter = log_format(base=8, mathtex=True)
+    assert formatter([1, 4, 64]) == ['$8^{0}$', '$8^{0.667}$', '$8^{2}$']
 
 
 def test_date_format():